compute/optimize/pyproject.toml
[project]
name = "mellivora-optimize"
version = "0.1.0"
description = "Optimization compute service for Mellivora Mind Studio"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.12.0",
    "cvxpy>=1.4.0",
    "structlog>=24.1.0",
    "pydantic>=2.5.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "nats-py>=2.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
